feat: to separate product update
